refactor: move playback/segments to playback reducer
