+++ src/store/reducers/playback/segments.js
